PITR: restore last chunk to the desired time
